requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.11.0

# Database
psycopg2-binary>=2.9.0
# sqlite3 is built-in to Python

# Data analysis
scikit-learn>=1.3.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Financial data
yfinance>=0.2.18
pandas-datareader>=0.10.0

# Backtesting
backtrader>=1.9.76
vectorbt>=0.25.0
empyrical>=0.5.5

# Web framework
flask>=2.3.0
flask-sqlalchemy>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0

# Data scraping
beautifulsoup4>=4.12.0
selenium>=4.11.0
lxml>=4.9.0

# Async and scheduling
aiohttp>=3.8.0
celery>=5.3.0
redis>=4.6.0

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
jupyter>=1.0.0
ipykernel>=6.25.0

# Optional: ML and advanced analytics
optuna>=3.2.0  # Hyperparameter optimization
ta-lib>=0.4.0  # Technical analysis (requires separate install)
